Move singularity and docker tests
